backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=9.5.0
pandas>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
aiofiles>=23.1.0